Updated Case Study model hierarchy
