+++ src/main/java/com/globallogic/amcr/service/impl/casestudies/CaseStudyServiceImpl.java
@@ -2,15 +2,14 @@ package com.globallogic.amcr.service.impl.casestudies;
 
 import com.globallogic.amcr.model.casestudies.CaseStudy;
 import com.globallogic.amcr.model.casestudies.CaseStudyOverview;
-import com.globallogic.amcr.model.casestudies.CaseStudyResponse;
 import com.globallogic.amcr.repository.casestudies.CaseStudyDao;
 import com.globallogic.amcr.service.casestudies.CaseStudyService;
 import com.globallogic.amcr.utils.Assert;
+import com.globallogic.amcr.utils.Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -41,21 +40,22 @@ public class CaseStudyServiceImpl implements CaseStudyService {
 
     @Override
     @Transactional(readOnly = true)
-    public CaseStudyResponse get(UUID id) {
+    public CaseStudy get(UUID id) {
         Assert.assertNotNull(id, "ID cannot be null to request entry");
         Log.debug("Service requesting case study with ID {}", id);
-        final CaseStudy caseStudy = caseStudyDao.get(id);
+        CaseStudy caseStudy = caseStudyDao.get(id);
         List<String> attachmentLinks = new ArrayList<>();
         if (caseStudy.getAttachmentIds() != null) {
             for (UUID attachmentId : caseStudy.getAttachmentIds()) {
-                attachmentLinks.add(ServletUriComponentsBuilder.fromCurrentRequest().path("/attachment/{id}").buildAndExpand(attachmentId).toUri().toString());
+                attachmentLinks.add(Utils.generateUri("/attachment/{id}", attachmentId).toString());
             }
         }
-        final String coverImageLink = ServletUriComponentsBuilder.fromCurrentRequest().path("/attachment/{id}").buildAndExpand(caseStudy.getCoverImageId()).toUri().toString();
         for (Map<String, String> row : caseStudy.getBody()) {
-            row.replace("imageId", ServletUriComponentsBuilder.fromCurrentRequest().path("/attachment/{id}").buildAndExpand(row.get("imageId")).toUri().toString());
+            row.replace("imageId", Utils.generateUri("/attachment/{id}", UUID.fromString(row.get("imageId"))).toString());
         }
-        return new CaseStudyResponse(caseStudy, attachmentLinks, coverImageLink);
+        caseStudy.setAttachmentLinks(attachmentLinks);
+        caseStudy.setCoverImageLink(Utils.generateUri("/attachment/{id}", caseStudy.getCoverImageId()).toString());
+        return caseStudy;
     }
 
     @Override
@@ -69,22 +69,35 @@ public class CaseStudyServiceImpl implements CaseStudyService {
     @Transactional
     public List<CaseStudyOverview> getAllOverviews() {
         Log.debug("Service requesting all case study overviews");
-        return caseStudyDao.getAllOverviews();
+        List<CaseStudyOverview> caseStudyOverviews = caseStudyDao.getAllOverviews();
+        for (CaseStudyOverview caseStudyOverview : caseStudyOverviews) {
+            caseStudyOverview.setCoverImageLink(Utils.generateUri("/attachment/{id}", caseStudyOverview.getCoverImageId()).toString());
+        }
+        return caseStudyOverviews;
     }
 
     @Override
     @Transactional
     public List<CaseStudyOverview> getSpotlitOverviews() {
         Log.debug("Service requesting all spotlit case study overviews");
-        return caseStudyDao.getSpotlitOverviews();
+        List<CaseStudyOverview> caseStudyOverviews = caseStudyDao.getSpotlitOverviews();
+        for (CaseStudyOverview caseStudyOverview : caseStudyOverviews) {
+            caseStudyOverview.setCoverImageLink(Utils.generateUri("/attachment/{id}", caseStudyOverview.getCoverImageId()).toString());
+        }
+        return caseStudyOverviews;
     }
 
     @Override
     @Transactional
     public List<CaseStudyOverview> getLatestOverviews(int entries) {
         Assert.assertNotNull(entries, "Entries cannot be null");
-        Log.debug("Service requesting 5 most recent overviews");
-        return caseStudyDao.getLatestOverviews(entries);
+        Log.debug("Service requesting {} most recent overviews", entries);
+
+        List<CaseStudyOverview> caseStudyOverviews = caseStudyDao.getLatestOverviews(entries);
+        for (CaseStudyOverview caseStudyOverview : caseStudyOverviews) {
+            caseStudyOverview.setCoverImageLink(Utils.generateUri("/attachment/{id}", caseStudyOverview.getCoverImageId()).toString());
+        }
+        return caseStudyOverviews;
     }
 
     @Override
